refactor: remove ratatui/termion; iocraft migration complete
Phase 3: delete the superseded ratatui frontend now that the iocraft frontend
is live across all five tabs.
- Delete app.rs, signal.rs, ui/main_screen.rs, ui/startup_screen.rs, and the
  five ui/tab_*.rs; remove ui::draw and all parking #[allow(dead_code)].
- Move the Signal unit tests into backend/history.rs (their new home).
- Drop the ratatui and termion dependencies; refresh Cargo.lock.
- Update CLAUDE.md: backend->frontend data plane, backend/ + ui/ module layout,
  iocraft + smol dependencies.
Zero dead code, zero clippy warnings; ./ci/test_full.sh green (MSRV, feature
matrix, CLI smoke, doc tests); no ratatui/termion/tokio in the tree.
