+++ Cargo.toml
@@ -30,8 +30,6 @@ serde = { version = "1", features = ["derive"] }
 si-scale = "0.3"
 
 # ui
-ratatui = { version = "0.30", default-features = false, features = ["termion"] }
-termion = "4"
 iocraft = "0.8"
 smol = "2"
