+++ src/backend/history.rs
@@ -1,10 +1,8 @@
 //! Signal ring buffer + signal history (backend-owned).
 //!
-//! `Signal<T>` is the metric ring buffer ported verbatim from the former
-//! `src/signal.rs` (which now re-exports this type during the migration and is
-//! deleted in Phase 3). `History` and `HistoryExt::get_or_default` were lifted
-//! from `app.rs` so the collector thread owns all history state and the
-//! frontend holds none (MIGRATION.md §4.1).
+//! `Signal<T>` is the metric ring buffer. `History` and
+//! `HistoryExt::get_or_default` give the collector thread sole ownership of all
+//! history state so the frontend holds none (MIGRATION.md §4.1).
 
 use std::collections::HashMap;
 
@@ -81,5 +79,56 @@ pub(crate) trait HistoryExt {
 impl HistoryExt for History {
     fn get_or_default(&self, key: &MetricKey) -> &Signal<f32> {
         self.get(key).unwrap_or(&DEFAULT_SIGNAL)
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_signal_basics_u32() {
+        let mut signal = Signal::<u32>::with_capacity(3, /* max */ 4);
+        assert_eq!(signal.peak, 0);
+        signal.push(1);
+        signal.push(2);
+        signal.push(3);
+
+        assert_eq!(signal.as_slice(), &[1, 2, 3]);
+        assert_eq!(signal.peak, 3);
+
+        signal.push(4);
+        assert_eq!(signal.as_slice(), &[2, 3, 4]);
+        assert_eq!(signal.peak, 4);
+
+        for _ in 0..10 {
+            signal.push(1);
+        }
+        signal.push(0);
+        assert_eq!(signal.as_slice(), &[1, 1, 0]);
+        assert_eq!(signal.peak, 4);
+    }
+
+    #[test]
+    fn test_signal_basics_f32() {
+        let mut signal = Signal::<f32>::with_capacity(3, /* max */ 4.0);
+        assert_eq!(signal.peak, 0.0);
+        signal.push(1.0);
+        signal.push(2.0);
+        signal.push(3.0);
+
+        assert_eq!(signal.as_slice(), &[1, 2, 3]);
+        assert_eq!(signal.peak, 3.0);
+
+        signal.push(4.0);
+        assert_eq!(signal.as_slice(), &[2, 3, 4]);
+        assert_eq!(signal.peak, 4.0);
+
+        for _ in 0..10 {
+            signal.push(1.0);
+        }
+        signal.push(0.0);
+        assert_eq!(signal.as_slice(), &[1, 1, 0]);
+        assert_eq!(signal.peak, 4.0);
     }
 }
