+++ src/lib.rs
@@ -292,10 +292,6 @@
 //! [MIT license]: http://opensource.org/licenses/MIT
 //! [asitop]: https://github.com/tlkh/asitop
 
-// The ratatui `App` state is superseded by the iocraft `backend` + `PumasApp`
-// data plane; it stays compiled (no behavior change) until Phase 3 deletes it.
-#[allow(dead_code)]
-mod app;
 mod backend;
 pub mod config;
 pub mod error;
@@ -303,7 +299,6 @@ mod metric_key;
 mod metrics;
 mod modules;
 pub mod monitor;
-mod signal;
 mod ui;
 mod units;
