+++ src/ui/components/title_bar.rs
@@ -1,8 +1,8 @@
-//! Top title bar (MIGRATION.md §9.1, ratatui `main_screen.rs:31-50`).
+//! Top title bar (MIGRATION.md §9.1, the original `main_screen.rs:31-50`).
 //!
 //! A single row: the program name (`Pumas v{version}`) on the left in the
 //! default color, and the machine description (`{brand} (cores: …)`) on the
-//! right in the accent color. ratatui overlays a left-aligned and a
+//! right in the accent color. The original overlays a left-aligned and a
 //! right-aligned paragraph on the same row; iocraft expresses this as one
 //! `Row` with `justify_content: SpaceBetween`.
